requirements.txt
discord.py>=2.3.0
aiosqlite>=0.19.0
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
imagehash>=4.3.1
aiohttp>=3.9.0
python-dotenv>=1.0.0
numpy>=1.24.0
